requirements.txt
openai
mcp
fastmcp
nest-asyncio
llama-api-client
playwright
pytest
gitingest

# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# Llama 4 integration
llama-cpp-python==0.2.20
transformers==4.36.0
torch==2.1.1
accelerate==0.25.0

# Browser automation
playwright==1.40.0
selenium==4.15.2

# Code analysis
ast==0.0.2
tree-sitter==0.20.4
pygments==2.17.2

# Document parsing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
unstructured==0.11.0
unstructured[pdf]==0.11.0
unstructured[docx]==0.11.0
langchain==0.1.0
langchain-text-splitters==0.0.1
tiktoken==0.5.2
reportlab==4.0.7

# Visual generation
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
pillow==10.1.0

# API integrations
requests==2.31.0
aiohttp==3.9.1
httpx==0.25.2

# Data processing
pandas==2.1.4
numpy==1.25.2

# Utilities
click==8.1.7
rich==13.7.0
tqdm==4.66.1
python-multipart==0.0.6

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0 

